@@ -10,10 +10,10 @@
         "source": "https://github.com/google/recaptcha"
     },
     "require": {
-        "php": ">=5.3.2"
+        "php": ">=5.5"
     },
     "require-dev": {
-        "phpunit/phpunit": "4.5.*"
+        "phpunit/phpunit": "^4.8"
     },
     "autoload": {
         "psr-4": {
