@@ -51,9 +51,8 @@ class Socket
 
         if ($this->handle != false && $errno === 0 && $errstr === '') {
             return $this->handle;
-        } else {
-            return false;
         }
+        return false;
     }
 
     /**
