@@ -1,6 +1,6 @@
 <?php
 
-/* An autoloader for ReCaptcha\Foo classes. This should be require()d
+/* An autoloader for ReCaptcha\Foo classes. This should be required()
  * by the user before attempting to instantiate any of the ReCaptcha
  * classes.
  */
