@@ -21,11 +21,11 @@ class Drupal8Post implements RequestMethod {
   /**
    * Submit the POST request with the specified parameters.
    *
-   * @param RequestParameters $params
-   *   Request parameters
+   * @param ReCaptcha\RequestParameters $params
+   *   Request parameters.
    *
    * @return string
-   *   Body of the reCAPTCHA response
+   *   Body of the reCAPTCHA response.
    */
   public function submit(RequestParameters $params) {
 
